Manage Azure subscription information

Subscription defaults to NULL, in which case the subscription is expected to be enabled in the azure CLI cache already. Appends and sets to default subscription, if provided. Errors out if no subscription is enabled.

az_account(subscription = NULL, ...)

az_account_list(...)

az_account_show(subscription = NULL, ...)

az_account_set(subscription, ...)

Arguments

subscription
Name or ID of the Azure subscription to which costs are billed. According to an upvoted answer on Stack Overflow, Azure subscription IDs need not be considered a secret or personal identifiable information (PII). However, depending your applicable context and policies, you may want to provide this argument as a secret.

To find out which subscriptions you are currently authorised to use, run print(az_account_list()).

...
Arguments passed on to az_cli_run

cmd

Command, required, optional, additional, extra parameters, as for processx::run() add parameters are reserved for the user to pass down additional arguments to the Azure CLI. extra parameters are reserved for internal use.

req

Command, required, optional, additional, extra parameters, as for processx::run() add parameters are reserved for the user to pass down additional arguments to the Azure CLI. extra parameters are reserved for internal use.

opt

Command, required, optional, additional, extra parameters, as for processx::run() add parameters are reserved for the user to pass down additional arguments to the Azure CLI. extra parameters are reserved for internal use.

add

Command, required, optional, additional, extra parameters, as for processx::run() add parameters are reserved for the user to pass down additional arguments to the Azure CLI. extra parameters are reserved for internal use.

echo_cmd

Whether to print the command to run to the screen.

echo

Whether to print the standard output and error to the screen. Note that the order of the standard output and error lines are not necessarily correct, as standard output is typically buffered.

Details

Subscriptions are kept in the local azure CLI cache, so you should not have to run this more than once. On GitHub Actions, the azure login action will already set up a subscription.

Functions

az_account_list: Get a list of subscriptions for the logged in account.

az_account_show: Get the details of a subscription. If the subscription isn't specified, shows the details of the default subscription.

az_account_set: Set a subscription to be the current active subscription.