Manage Azure CLI configuration

Overwrites defaults for the Azure CLI to .azure/ directory (side effect), if arguments are provided. Errors out if a default is missing.

az_configure(name = NULL, resource_group = NULL, ...)

az_configure_list(...)

Arguments

name
Name of the web app. (In the Azure CLI, this argument is sometimes known as name, and sometimes as web).

resource_group
The Azure resource group to which the app service should belong.

...
Arguments passed on to az_cli_run

cmd

Command, required, optional, additional, extra parameters, as for processx::run() add parameters are reserved for the user to pass down additional arguments to the Azure CLI. extra parameters are reserved for internal use.

req

Command, required, optional, additional, extra parameters, as for processx::run() add parameters are reserved for the user to pass down additional arguments to the Azure CLI. extra parameters are reserved for internal use.

opt

Command, required, optional, additional, extra parameters, as for processx::run() add parameters are reserved for the user to pass down additional arguments to the Azure CLI. extra parameters are reserved for internal use.

add

Command, required, optional, additional, extra parameters, as for processx::run() add parameters are reserved for the user to pass down additional arguments to the Azure CLI. extra parameters are reserved for internal use.

echo_cmd

Whether to print the command to run to the screen.

echo

Whether to print the standard output and error to the screen. Note that the order of the standard output and error lines are not necessarily correct, as standard output is typically buffered.

Details

Because of a bug, the Azure CLI will always include defaults at ~/.azure/. These hidden defaults can interfere with these functions. Make sure that you have no default name and resource_group in the global azure default config in your HOME directory.

Functions

az_configure_list: List defaults