Log in to Azure

Helper for interactive login.

az_login(...)

Arguments

...
Arguments passed on to az_cli_run

cmd

Command, required, optional, additional, extra parameters, as for processx::run() add parameters are reserved for the user to pass down additional arguments to the Azure CLI. extra parameters are reserved for internal use.

req

Command, required, optional, additional, extra parameters, as for processx::run() add parameters are reserved for the user to pass down additional arguments to the Azure CLI. extra parameters are reserved for internal use.

opt

Command, required, optional, additional, extra parameters, as for processx::run() add parameters are reserved for the user to pass down additional arguments to the Azure CLI. extra parameters are reserved for internal use.

add

Command, required, optional, additional, extra parameters, as for processx::run() add parameters are reserved for the user to pass down additional arguments to the Azure CLI. extra parameters are reserved for internal use.

echo_cmd

Whether to print the command to run to the screen.

echo

Whether to print the standard output and error to the screen. Note that the order of the standard output and error lines are not necessarily correct, as standard output is typically buffered.